requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==26.0.0
httpie==3.2.2
idna==3.7
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
